divided into grid
