+++ src/components/TaskBoard/index.scss
@@ -1,7 +1,6 @@
 .task-board {
     margin: auto;
     max-width: 78rem;
-    justify-content: space-between;
 
     &__container {
         gap: 2.125rem;
